refactor: moved search and sort to rout
