+++ youtube-client-app/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { SearchResultsComponent } from './components/search-results/search-results.component';
+import { SearchResultsComponent } from './youtube/components/search-results/search-results.component';
 import { ErrorComponent } from './core/pages/error/error.component';
 
 const routes: Routes = [
