+++ youtube-client-app/src/app/app.module.ts
@@ -5,26 +5,28 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app/app.component';
 import { AuthModule } from './auth/auth.module';
-import { SearchItemComponent } from './components/search-item/search-item.component';
-import { SearchResultsComponent } from './components/search-results/search-results.component';
-import { SortByNamePipe } from './components/search-results/sort-by-name.pipe';
+import { SearchItemComponent } from './youtube/components/search-item/search-item.component';
+import { SearchResultsComponent } from './youtube/components/search-results/search-results.component';
+import { SortByNamePipe } from './youtube/components/search-results/sort-by-name.pipe';
 import { CoreModule } from './core/core.module';
 import { MaterialModule } from './material.module';
+import { YoutubeModule } from './youtube/youtube.module';
 
 @NgModule({
   declarations: [
-	AppComponent,
-	SearchResultsComponent,
-	SearchItemComponent,
-	SortByNamePipe,
+    AppComponent,
+    SearchResultsComponent,
+    SearchItemComponent,
+    SortByNamePipe,
   ],
   imports: [
-	BrowserModule,
-	AppRoutingModule,
-	BrowserAnimationsModule,
-	MaterialModule,
-  AuthModule,
-  CoreModule,
+    BrowserModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    MaterialModule,
+    AuthModule,
+    CoreModule,
+    YoutubeModule,
   ],
   providers: [],
   bootstrap: [AppComponent],
