+++ youtube-client-app/src/app/app/app.component.html
@@ -1,2 +1,2 @@
-<app-header (getVideoName)="getName($event)" (sb)="sortBy($event)" (sbw)="sortByWord($event)"></app-header>
+<app-header></app-header>
 <router-outlet></router-outlet>
