+++ youtube-client-app/src/app/app/app.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -6,19 +6,5 @@ import { ChangeDetectorRef, Component } from '@angular/core';
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-  public videoName: string;
-  public sort: string;
-  public sortWord: string;
 
-  public getName(name: string): void {
-  this.videoName = name;
-  }
-
-  public sortBy(type: string): void {
-  this.sort = type;
-  }
-
-  public sortByWord(name: string): void {
-  this.sortWord = name;
-  }
 }
