+++ youtube-client-app/src/app/core/components/header/header.component.html
@@ -1,5 +1,5 @@
 <div class="header">
   <div class="header-item"></div>
-  <app-search-panel (getName)="getName($event)" (sb)="sortBy($event)" (sbw)="sortByWord($event)"></app-search-panel>
+  <app-search-panel></app-search-panel>
   <app-user-login></app-user-login>
 </div>
