+++ youtube-client-app/src/app/core/components/header/header.component.scss
@@ -9,12 +9,4 @@
   &-item {
     flex: 1;
   }
-
-  app-search-panel {
-    flex: 2;
-  }
-
-  app-user-login {
-    flex: 1;
-  }
 }
