+++ youtube-client-app/src/app/core/components/search-panel/search-panel.component.scss
@@ -1,10 +1,10 @@
+:host {
+  flex: 2;
+}
+
 .container {
   display: flex;
   flex-direction: column;
-}
-
-app-sort-panel {
-  overflow: hidden;
 }
 
 .search-panel,
