+++ youtube-client-app/src/app/core/components/sort-panel/sort-panel.component.html
@@ -4,18 +4,18 @@
     <p class="sort-panel__title">Sorting by:</p>
 
     <ul class="sort-panel__list">
-      <li (click)="sortBy('date')" class="sort-panel__item selected">
+      <li (click)="sortBy('date')" [ngClass]="{selected: isDate}" class="sort-panel__item">
         date
       </li>
-      <li (click)="sortBy('view')" class="sort-panel__item selected">
+      <li (click)="sortBy('view')" [ngClass]="{selected: isView}" class="sort-panel__item">
         count of view
       </li>
-      <li (click)="sortByWord(input)" class="sort-panel__item">
+      <li (click)="sortByWord($event, input)" [ngClass]="{selected: isWord}" class="sort-panel__item">
         by word or sentence
       </li>
       <li class="sort-panel__item">
         <mat-form-field class="sort-panel__input">
-          <input #input matInput placeholder="Word or sentence">
+          <input (keypress)="sortByWord($event, input)" #input matInput placeholder="Word or sentence">
         </mat-form-field>
       </li>
     </ul>
