+++ youtube-client-app/src/app/core/components/sort-panel/sort-panel.component.scss
@@ -1,5 +1,9 @@
 @import "../../../../styles/vars";
 
+:host {
+  overflow: hidden;
+}
+
 .sort-panel {
   display: flex;
   align-items: center;
