+++ youtube-client-app/src/app/core/components/user-login/user-login.component.scss
@@ -1,3 +1,7 @@
+:host {
+  flex: 1;
+}
+
 .user-login {
   height: 100%;
   display: flex;
