+++ youtube-client-app/src/app/core/pages/error/error.component.html
@@ -1,6 +1,4 @@
-<div class="error-page">
-  <div class="error-page__wrap">
-    <img class="error-page__img" src="assets/img/404.svg" alt="error">
-    <p class="error-page__text">Sorry, smth went wrong((</p>
-  </div>
+<div class="error-page__wrap">
+  <img class="error-page__img" src="assets/img/404.svg" alt="error">
+  <p class="error-page__text">Sorry, smth went wrong((</p>
 </div>
