+++ youtube-client-app/src/app/core/services/data.service.ts
@@ -1,10 +1,10 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class DataService {
-
-  constructor() { }
-
-}
+// import { Injectable } from '@angular/core';
+//
+// @Injectable({
+//   providedIn: 'root',
+// })
+// export class DataService {
+//
+//   constructor() { }
+//
+// }
