+++ youtube-client-app/src/app/youtube/components/search-item/search-item.component.html
@@ -10,19 +10,19 @@
       <ul class="card__info-list">
 
         <li class="card__info-item">
-          <img src="assets/img/viewed.svg" alt="view">
+          <img src="../../../../assets/img/viewed.svg" alt="view">
           <span class="card__info-num">{{statistics.viewCount}}</span>
         </li>
         <li class="card__info-item">
-          <img src="assets/img/liked.svg" alt="like">
+          <img src="../../../../assets/img/liked.svg" alt="like">
           <span class="card__info-num">{{statistics.likeCount}}</span>
         </li>
         <li class="card__info-item">
-          <img src="assets/img/dislike.svg" alt="dislike">
+          <img src="../../../../assets/img/dislike.svg" alt="dislike">
           <span class="card__info-num">{{statistics.dislikeCount}}</span>
         </li>
         <li class="card__info-item">
-          <img src="assets/img/comments.svg" alt="comments">
+          <img src="../../../../assets/img/comments.svg" alt="comments">
           <span class="card__info-num">{{statistics.commentCount}}</span>
         </li>
