+++ youtube-client-app/src/app/youtube/components/search-item/search-item.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/vars";
+@import "../../../../styles/vars";
 
 .search-item {
   padding: 14px;
