+++ youtube-client-app/src/app/youtube/components/search-results/search-results.component.scss
