+++ youtube-client-app/src/app/youtube/youtube.module.ts
@@ -1,0 +1,10 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+  ],
+})
+export class YoutubeModule { }
